+++ src/operations/Completion.cpp
@@ -542,9 +542,8 @@ std::vector<lsp::CompletionItem> WorkspaceFolder::completion(const lsp::Completi
                 }
 
                 // Check if it starts with a directory alias, otherwise resolve with require base path
-                std::filesystem::path currentDirectory =
-                    resolveDirectoryAlias(rootUri.fsPath(), config.require.directoryAliases, contentsString, /* includeExtension = */ false)
-                        .value_or(fileResolver.getRequireBasePath(moduleName).append(contentsString));
+                std::filesystem::path currentDirectory = resolveDirectoryAlias(rootUri.fsPath(), config.require.directoryAliases, contentsString)
+                                                             .value_or(fileResolver.getRequireBasePath(moduleName).append(contentsString));
 
                 try
                 {